+++ CHANGELOG.markdown
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Update dependencies. (Tristan Dunn)
+
 ## 0.5.0 — January 21, 2013
 
 * Support channel, channels, and user REST API endpoints. (Tristan Dunn)