+++ CHANGELOG.markdown
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.6.0 — January 23, 2013
 
 * Add a file for easily starting the fake server in Cucumber. (Tristan Dunn)
 * Add convenience method for the JS to override Pusher client configuration. (Tristan Dunn)